Split unit test data to service specific parts.
